.emacs: kill irony-mode
Although I haven't worked with C++ much for the past year to judge it
myself, but with pretty much every other lang I use lsp-mode now, and
judging by comments on reddit for "lsp-mode vs irony" it is just as good
for C++. So should be no reason to keep irony. On of that, it seems
irony become unmaintained (I didn't check myself, but people on reddit
say so).

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>