Revert ".emacs: use is-in-comment detection by smartparens"
This reverts commit 48a2ae8bec3c842ed098e810ea221da2ff33cd0b.

First of all, heir detection is broken. If I put a caret inside braces
in a comment in python-mode, their function will say I am not in comment
(for some reason it thinks — or more specifically it is what `context`
variable is set to — I am in a string).

In addition to that, moving the handler to :unless has its own problem
in that braces autocompletion stops working completely while in a comment
(look, now it detected comment-context correctly. Too bad for me.)