Revert ".gitconfig: show diffs right away if it fits the screen"
This reverts commit ce568e641aec0261c9fd847b6acbe73bfd906487.

Actual user-experience shows, having diffs always buffered is more
comfortable. That is because I may want to see `git diff`, `git log`,
etc, and a pile of diffs will be assembling on the screen. And then, it
becomes hard to visually differentiate where one diff ends and another
one starts. You need to look for the cmd prompt that separates them.
While buffering the output, the problem doesn't exist.