.emacs: use cl-assert on :config stage
There's little reason for the assert to be at all. AFAIR the reason it
is there is to avoid warnings on byte-compilation. Other than that, if I
won't get completions, I'll realize something is wrong with company-mode.

So anyway, let's move the assert to :config stage, for its purposes it
should be enough.