prepare-commit-msg: add heuristics for repeating prev. title
Sometimes I do changes to same files multiple commits in a row, and I
don't want to rewrite the title over again. Let's add a simple heuristic
that if same file was changed but title is different, that means a user
changed it, let's do the same for the new commit.