.emacs: disable "ensure" for uninstallable packages
I will need it in order to be able to download latest packages on a new
system. Per documentation, I should've also added a

    (require 'use-package-ensure)
    (setq use-package-always-ensure t)
…but it is buggy. It always connects ELPA/MELPA/etc, disregarding
whether packages are installed or not. It's funny, given it can't
autoupdate (well, not without some other setup).

No point in fixing the bug though: use-package seems unmaintained.
There's a pile of PRs, but nobody merges them in.