.emacs: do not set up git-commit, set up magit instead
First it works then it doesn't… Apparently, git-commit is a part of
magit project, and I had to move all setup to magit instead. At this
point I just hope that the "it works" now really works, but not "work
now but will break tomorrow". Restarting Emacs a bunch of times,
everything seem fine.