.emacs: make sure magit-auto-revert mode is down
Now I'm at the point of enough annoyance to call magit a piece of
shit. This time I stumbled upon this garbage being enabled, which I
never wanted and explicitly don't want, because I use instead
auto-revert mode, which is much better because it does not litter with
excess IO the files, which matters when a file you're visiting is
located on a network-mounted dir.