.emacs: restore default company-backends
Apparently, I read the situation the wrong way. The `company-capf` is
the go to backend in many situations, in particular with lsp-mode and
clangd. And it does not lag in that particular situation. So I gotta
say, apparently it depends on the backend to the backend, and whenever
`company-capf` happens to lag, that needs to be investigated
separately.