.vimrc: drop a few limitations for overly big files
I'm not sure these give much of a win. I mean, if you don't use undo
buffer, it shouldn't take any noticeable size, right? So what's the
point in disabling it? 🤷‍♂️ Initially the code was part of someone's
else config (likely posted on SO), so I just copied it over and never
really gave it much thought. But now that I think of it, disabling
write functionality and undo-buffer seems like an overkill for no
reason.

For what it worth though, I did not test my claim that these features
being on do not take much memory. It just seems kinda rational.