.emacs: highlight-parentheses: change colors order
It kinda seems more logical: green is the closest, yellow is the
second from closes, as in traffic lights. And then I use "orange" as
it's something in between "yellow" and "red". And then goes "pink",
because pure "red" would mix up with "error highlight" which some
modes may apply to misplaced braces, and pink is the next close enough
color that you wouldn't mix up with red.