.emacs: improve projectile-find-file-from-primary-clipboard
Older version didn't support finding a file without a path to it. This
one works with plain filenames as well.

Projectile is no longer used to find the file, but I've left the name
for now as the function still assumed to be used with projects. Maybe
some day I'll rewrite it to make full use of projectile, although it's
unlikely, don't see a usecase coming up.