.emacs: move "melpa" addition to be before any other `use-package`
Apparently, the problems I had with some packages not getting
installed is simply because I had melpa added to late in the config.

While at it, remove a comment about a "bug". I do not remember the
details why it was added, but I would presume the problem being fixed
here is exactly the reason.

To be fair, the error about `foo-` not being available when `foo` is
being asked is terrible, so no wonder I considered it a bug. But it
turns out to be just a bad debugging print, the dash at the end isn't
actually being added when Emacs tries to install package `foo`, only
when it prints the error.