.emacs: find-file-from-primary-clipboard: support `file.c:12:34` syntax
Such syntax is used by GCC in their warnings/errors, so that's useful
to support. For simplicity the code assumes for now a file/path may not
contain a colon. The possibility of not only encountering such file
but also having to call this function on it is so low that I just
couldn't be bothered to work on that.