.emacs: fonts: don't duplicate :height and equalize it
It's the same as one in 'default, so should simply be inherited.
Except the one that has 95 size, however I've been using same size for
a long time and didn't even know the dotfiles had different size. I
guess that means there's no practical reason for the sizes to be
different, so make it equal to everything else.