.emacs: "markdown to bbcode": replace `hX` with `SIZE` + `B` tags
It seems, `h1`, etc, isn't a thing in BBCode, or at least in Bitrix
version of it (but I didn't find in general format description such
support as well).

However, Bitrix BBCode supports `SIZE`, which may be used as a
workaround to render headings.

The "em" sizes are taken from Github rendering of headings in Firefox.